moved Shape interface outside shapes package (implementations only); increased encapsulation for Rectangle via field access levels and getters/setters
