+++ java/src/main/java/edu/cmu/cs/cs214/lab02/Renderer.java
@@ -1,7 +1,5 @@
 package edu.cmu.cs.cs214.lab02;
 
-import edu.cmu.cs.cs214.lab02.shapes.*;
-
 public class Renderer {
     public Shape shape;
