+++ java/src/main/java/edu/cmu/cs/cs214/lab02/shapes/Circle.java
@@ -1,5 +1,7 @@
 package edu.cmu.cs.cs214.lab02.shapes;
 
+import edu.cmu.cs.cs214.lab02.Shape;
+
 public class Circle implements Shape {
     public double radius;
